refactor: mv rebase options to dedicated modules
